refactor(agentic-ai): re-add tools path to cache/retries configuration entry
