+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/autoconfigure/AgenticAiConnectorsConfigurationProperties.java
@@ -17,18 +17,23 @@ import org.springframework.validation.annotation.Validated;
 @Validated
 @ConfigurationProperties(prefix = "camunda.connector.agenticai")
 public record AgenticAiConnectorsConfigurationProperties(
-    @Valid @NotNull @DefaultValue ProcessDefinitionConfiguration processDefinition) {
-  public record ProcessDefinitionConfiguration(
-      @Valid @NotNull @DefaultValue RetriesConfiguration retries,
-      @Valid @NotNull @DefaultValue CacheConfiguration cache) {
+    @Valid @NotNull @DefaultValue ToolsConfiguration tools) {
 
-    public record RetriesConfiguration(
-        @DefaultValue("4") @PositiveOrZero Integer maxRetries,
-        @DefaultValue("PT0.5S") Duration initialRetryDelay) {}
+  public record ToolsConfiguration(
+      @Valid @NotNull @DefaultValue ProcessDefinitionConfiguration processDefinition) {
 
-    public record CacheConfiguration(
-        @DefaultValue("true") boolean enabled,
-        @DefaultValue("100") @PositiveOrZero Long maximumSize,
-        @DefaultValue("PT10M") Duration expireAfterWrite) {}
+    public record ProcessDefinitionConfiguration(
+        @Valid @NotNull @DefaultValue RetriesConfiguration retries,
+        @Valid @NotNull @DefaultValue CacheConfiguration cache) {
+
+      public record RetriesConfiguration(
+          @DefaultValue("4") @PositiveOrZero Integer maxRetries,
+          @DefaultValue("PT0.5S") Duration initialRetryDelay) {}
+
+      public record CacheConfiguration(
+          @DefaultValue("true") boolean enabled,
+          @DefaultValue("100") @PositiveOrZero Long maximumSize,
+          @DefaultValue("PT10M") Duration expireAfterWrite) {}
+    }
   }
 }
