+++ connectors/agentic-ai/src/test/java/io/camunda/connector/agenticai/autoconfigure/AgenticAiConnectorsAutoConfigurationTest.java
@@ -155,7 +155,8 @@ class AgenticAiConnectorsAutoConfigurationTest {
   @Test
   void whenToolsCachingDisabled_thenConfiguresDefaultToolElementsResolver() {
     contextRunner
-        .withPropertyValues("camunda.connector.agenticai.process-definition.cache.enabled=false")
+        .withPropertyValues(
+            "camunda.connector.agenticai.tools.process-definition.cache.enabled=false")
         .run(
             context ->
                 assertThat(context)
@@ -166,7 +167,8 @@ class AgenticAiConnectorsAutoConfigurationTest {
   @Test
   void whenToolsCachingEnabled_thenConfiguresCachingToolElementsResolver() {
     contextRunner
-        .withPropertyValues("camunda.connector.agenticai.process-definition.cache.enabled=true")
+        .withPropertyValues(
+            "camunda.connector.agenticai.tools.process-definition.cache.enabled=true")
         .run(
             context ->
                 assertThat(context)
@@ -177,7 +179,8 @@ class AgenticAiConnectorsAutoConfigurationTest {
   @Test
   void whenToolsCachingMaximumSizeIsNegative_thenFailsValidation() {
     contextRunner
-        .withPropertyValues("camunda.connector.agenticai.process-definition.cache.maximumSize=-10")
+        .withPropertyValues(
+            "camunda.connector.agenticai.tools.process-definition.cache.maximum-size=-10")
         .run(
             context ->
                 assertThat(context)
@@ -198,7 +201,7 @@ class AgenticAiConnectorsAutoConfigurationTest {
                                   FieldError::getDefaultMessage)
                               .containsExactly(
                                   "camunda.connector.agenticai",
-                                  "processDefinition.cache.maximumSize",
+                                  "tools.processDefinition.cache.maximumSize",
                                   -10L,
                                   "must be greater than or equal to 0");
                         }));
